+++ software/examples/csharp/ExampleThreshold.cs
@@ -9,7 +9,7 @@ class Example
 	private static string UID = "XYZ"; // Change XYZ to the UID of your Laser Range Finder Bricklet 2.0
 
 	// Callback function for distance callback
-	static void DistanceCB(BrickletLaserRangeFinderV2 sender, int distance)
+	static void DistanceCB(BrickletLaserRangeFinderV2 sender, short distance)
 	{
 		Console.WriteLine("Distance: " + distance + " cm");
 	}
